% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Login to iAsset and retreive access token}
\usage{
get_access_token(username, quiet = FALSE)
}
\arguments{
\item{username}{the iAsset username}

\item{quiet}{Should the function return verbose messages}
}
\value{
Invisibily, an access token upon succes
}
\description{
Login to iAsset and retreive access token
}
