% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get information about a custom inspection}
\usage{
get_fields(name = "Vespa-Watch", access_token = get_access_token(quiet = TRUE))
}
\arguments{
\item{name}{name of the custom inspection to return fields for}

\item{access_token}{access token from \code{get_access_token()}}
}
\value{
a list object with the name, id and fields (data.frame) of the custom
inspection
}
\description{
Get information about a custom inspection
}
\examples{
\dontrun{get_fields("Vespa-Watch")}
}
