% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Title}
\usage{
get_records(
  inspection_name = "Vespa-Watch",
  access_token = get_access_token(quiet = TRUE)
)
}
\arguments{
\item{inspection_name}{name of the custom inspection to return records for}

\item{access_token}{access token from \code{get_access_token()}}
}
\value{
a tibble with the records from the selected inspection.
}
\description{
Title
}
\examples{
\dontrun{get_records("Vespa-Watch")}
}
