% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_username.R
\name{get_username}
\alias{get_username}
\title{Get the username associated with the iAsset credentials in the system credentials store.}
\usage{
get_username(service = "iasset_password")
}
\arguments{
\item{service}{Character. The keyring service to fetch the stored username for.}
}
\value{
A character vector of stored usernames for the service. NA if none.
}
\description{
Get the username associated with the iAsset credentials in the system credentials store.
}
\examples{
\dontrun{get_username("iasset_password")}
}
