% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iassetR-package.R
\docType{package}
\name{iassetR-package}
\alias{iassetR}
\alias{iassetR-package}
\title{iassetR: Use the iAsset API in R}
\description{
This package is an interface to the REST api of iAsset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inbo.github.io/iassetR/}
  \item Report bugs at \url{https://github.com/inbo/iassetR/issues}
}

}
\author{
\strong{Maintainer}: Sander Devisscher \email{sander.devisscher@inbo.be} (\href{https://orcid.org/0000-0003-2015-5731}{ORCID})

Authors:
\itemize{
  \item Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})
  \item Diego Neve \email{diegoneve.26@gmail.com} (\href{https://orcid.org/0009-0003-2454-8521}{ORCID})
}

}
\keyword{internal}
