% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get all records in a custom inspection}
\usage{
get_records(
  inspection_name = "Vespa-Watch",
  access_token = get_access_token(quiet = TRUE),
  get_metadata = c("id", "none", "all")
)
}
\arguments{
\item{inspection_name}{name of the custom inspection to return records for}

\item{access_token}{access token from \code{get_access_token()}}

\item{get_metadata}{Indicates whether to include metadata in the returned
data. Possible values are "id" (only include observation \code{insp_order}),
"none" (exclude all metadata), and "all" (include all available
metadata). Default is "id".}
}
\value{
a tibble with the records from the selected inspection.
}
\description{
Get all records in a custom inspection
}
\examples{
\dontrun{get_records("Vespa-Watch")}
}
