% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iassetR-package.R
\docType{package}
\name{iassetR-package}
\alias{iassetR}
\alias{iassetR-package}
\title{iassetR: Use the iAsset API in R.}
\description{

}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})

Authors:
\itemize{
  \item Sander Devisscher \email{sander.devisscher@inbo.be} (\href{https://orcid.org/0000-0003-2015-5731}{ORCID})
}

}
\keyword{internal}
